BAEL-2394 JPA One-to-One Relationship
