+++ persistence-modules/hibernate5/src/main/java/com/baeldung/hibernate/onetoone/HibernateUtil.java
@@ -1,0 +1,66 @@
+package com.baeldung.hibernate.onetoone;
+
+import com.baeldung.hibernate.customtypes.LocalDateStringType;
+import org.hibernate.SessionFactory;
+import org.hibernate.boot.Metadata;
+import org.hibernate.boot.MetadataSources;
+import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
+import org.hibernate.service.ServiceRegistry;
+
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.net.URL;
+import java.util.Properties;
+
+public class HibernateUtil {
+    private static SessionFactory sessionFactory;
+
+    private HibernateUtil() {
+    }
+
+    public static SessionFactory getSessionFactory(Strategy strategy) {
+        if (sessionFactory == null) {
+            sessionFactory = buildSessionFactory(strategy);
+        }
+        return sessionFactory;
+    }
+
+    private static SessionFactory buildSessionFactory(Strategy strategy) {
+        try {
+            ServiceRegistry serviceRegistry = configureServiceRegistry();
+
+            MetadataSources metadataSources = new MetadataSources(serviceRegistry);
+
+            for (Class<?> entityClass : strategy.getEntityClasses()) {
+                metadataSources.addAnnotatedClass(entityClass);
+            }
+
+            Metadata metadata = metadataSources.getMetadataBuilder()
+                    .applyBasicType(LocalDateStringType.INSTANCE)
+                    .build();
+
+            return metadata.getSessionFactoryBuilder()
+                    .build();
+        } catch (IOException ex) {
+            throw new ExceptionInInitializerError(ex);
+        }
+    }
+
+
+    private static ServiceRegistry configureServiceRegistry() throws IOException {
+        Properties properties = getProperties();
+        return new StandardServiceRegistryBuilder().applySettings(properties)
+                .build();
+    }
+
+    private static Properties getProperties() throws IOException {
+        Properties properties = new Properties();
+        URL propertiesURL = Thread.currentThread()
+                .getContextClassLoader()
+                .getResource("hibernate.properties");
+        try (FileInputStream inputStream = new FileInputStream(propertiesURL.getFile())) {
+            properties.load(inputStream);
+        }
+        return properties;
+    }
+}
