+++ persistence-modules/hibernate5/src/main/java/com/baeldung/hibernate/onetoone/foreignkeybased/User.java
@@ -22,7 +22,7 @@ public class User {
     private String userName;
 
     @OneToOne(cascade = CascadeType.ALL)
-     @JoinColumn(name = "address_id", referencedColumnName = "id")
+    @JoinColumn(name = "address_id", referencedColumnName = "id")
     private Address address;
 
     public Long getId() {
